feat: finalize code for frontend and backend

+++ backend/src/main/java/com/stp/straightthroughprocessing/model/User.java
@@ -16,6 +16,6 @@ public class User {
     private String name;
 
     @Column(name = "cash")
-    private Long cash;
+    private Double cash;
 }
