+++ backend/src/main/java/com/stp/straightthroughprocessing/model/request/ModifyBalanceUser.java
@@ -2,5 +2,5 @@ package com.stp.straightthroughprocessing.model.request;
 
 public class ModifyBalanceUser {
     public Long id;
-    public Long amount;
+    public Double amount;
 }
